Remove dbus/gnome-keyring configuration
It doesn't make sense for these configurations to be tied to openbox.
`dbus` is launched by systemctl.  It has to be informed about the X
server, so that gnome-keyring can use it to open pinentry GUIs, but this
is done in `.xinitrc`.  `gnome-keyring` is launched by PAM, so that it
can be automatically unlocked.